pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-research-cli"
version = "0.1.0"
description = "Backend-first CLI trading research system for time-series model training and evaluation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
    "pydantic>=2.10.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "typer>=0.15.1",
    "matplotlib>=3.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
